removed BigNumberJS, fixed precision error
The BigNumber variable was removed and all BigNumbers are now gotten through MathJS. This wasn't the cause of the precision error, but MathJS did everything BigNumberJS did also, so messing with BigNumberJS was making things more complicated than they needed to be.

The precision error was fixed by turning the math.evaluate() to use bignumber x-values in the getXSteps function.