Added 3d graphing functionality
Users can now graph, save, and recall points for 3d functions.

A new todo has been added for the 3d functionality. It's a small problem, but is clear for functions like x^2 + y^2. It's usually not noticeable, though.

Broke up some of the logic for getting points into separate methods.