Fixed todos, added auto recalculations for 2D graphs.
PlotlyJS was also fixed. There was a temp fix in TSConfig for it, but the new import in the app.module and component should be a better solution.

Some TODOs were added for the changing window and auto recalculate functions. One is to make it work for 3D, which will be done soon.

Additionally, some HTML was added to tell the user how many points were found in the DB and how many were needed to be calculated on their end. This is to help visually show the usefulness for this program.