Fixed filtering Points, added quicker sort, unit testing
Added a sorting function that gets around the .some() being called thousands of times. This helped the program speed up quite a bit for 3D graphs.

Additionally, an issue with filtering points was solved. Now it shouldn't be ignoring some points that do belong, causing the program to recalculate points unnecessarily.

The sorting feature also allows for duplicate points to be ignored. This ensures that if a bug does cause duplicate points in the DB, it won't affect the graph on the front-end.

There was also a checkForProblems function that is essentially unit testing. It complains if something is wrong with the graph and helps point out where that issue is arising.