Added input validation and error messages
There's some validation done before getting the graph. It checks if there's an equation currently entered, valid step values, and windows.

Some of these still need to be fleshed out a bit more, but when one is triggered, the graph does not get calculated. It also displays a red message above where the error is with info on what is wrong.