Added theme slider, removed some todos.
I added some extra CSS to change the theme of the website and make the buttons look a bit better. This allows the user to have light/dark mode.

Some todos were removed. These were already finished, but the comments were not removed.

I also removed many of the booleans from the program used to display tags on the front-end. These cluttered the code. I instead removed them all and used DOM to display or not display these sections. It made the validateData() function a bit harder to read, but overall not having dozens of booleans that exists only for one function is nice.