Optimized back-end
The Points table was removed from the README and now the user needs to just have the Grapher DB and connection string. The program now creates, adds, and retrieves from table at runtime automatically.

Prior to this change, points were retrieved by looking at every row of Points, a table that consisted of all points, and checking if it had the needed equation. This was quite expensive as it required the program to potentially search through millions of points.

The new tables now are named after the equation that they draw their points from. This cuts out the need to search through every point ever found, which is overall a time save.

Because point objects no longer need to store equation info, that was also removed to save memory. Id for the points were also removed.

A large amount of the DBContext file was also deleted. Since the Points table is no longer used, those lines were removed.