Updated README
Small update to explain why users should have to run SQL commands that they don't necessarily need to be ran by them. It feels weird to me to create a database on someone's server without telling them and then filling it with a bunch of data that'll eat up memory.

I wouldn't want a small program that's supposed to be a fun niche thing to eat up gigs without tell me, so I'm not going to do that to others.

This way, if the user runs the commands and forgets to delete the database after no longer using the program and have their memory eaten, then that's their fault.