Fixed table naming issue
Equations used are now all saved to a central table, which has a name corresponding to the table that holds the points for the used equation. There's still a few issues that need to be worked on with this, but the program is functioning and the Equations table is made automatically.