Added count points and clear table buttons.
These are very simple and will be changed soon. The buttons allow the user to query the DB about how many points are stored for a particular equation. The clear table button is to allow user to delete all points for a particular equation.

An issue with the program is that if tables grow too large, SQL queries and filter functions will take longer than simply calculating points. This is because it requires a number of calculations in order to determine if a point belongs in the graph or not. For very complicated equations, the number of points that can be stored and still worthwhile to search for is rather large, but simpler equations require tables only store a small number of points for any potential time-saving.

For this reason, I felt it worthwhile to add a feature that allows users to clear a table.