All equations with stored points now shown
Users now have a list of all equations in the DB that contain stored points. These have a clear points button next to them that allows the user to delete all points in the associated table.

The reasoning behind the button is that querying and filtering points can take
significantly more time than simply letting the processor calculate all the points when the table is sufficiently large. The button lets users clear data in tables that have grown too large and are no longer worthwhile to use.

There's still a bit more that needs to be done in regards to this addition, but in the future I hope to implement some feature that detects when a table has become too large and should be deleted when considering the difficulty of its equation.