Refactored back end a bit
The program now doesn't need to constantly call the DB to find table names. Additionally, the front end also no longer needs to query the DB to get basic data. The program got around this by creating lists on its end and sending data back to the DB. The list holds all data we will need when it comes to table names and equations and whatnot.