Added button to graph known equations.
Users can now click a button that retrieves all points for an equation, then graphs them. This is useful for users to visualize what the known data looks like and decide if the table should be cleared, because it might have data that won't be in their window settings, which would be a waste to filter through.