Fixed some todos.
There were a few small fixes for TODOs. The program's development will be ended soon. The idea behind it, pulling points in from the DB, isn't all that useful when actively looking around the same equation. It takes time and almost certainly the points retrieved aren't a valid point to be used, so it's just a waste of calculations for the computer.